+++ Frontend/src/App.js
@@ -3,6 +3,7 @@ import "./sass/main.scss";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Main from "../src/pages/landing page/Main";
 import ResetPassword from "../src/pages/reset password/Main";
+import ComponentShowcase from "./pages/component-showcase/Main";
 
 const App = () => {
   return (
@@ -10,6 +11,7 @@ const App = () => {
       <Routes>
         <Route exact path="/" element={<Main />}></Route>
         <Route path="/reset password" element={<ResetPassword />}></Route>
+        <Route path="/component" element={<ComponentShowcase />}></Route>
       </Routes>
     </BrowserRouter>
   );
