+++ Frontend/src/static/components/icons/apple-icon.jsx
