+++ Frontend/src/static/components/icons/email-icon.jsx
