+++ Frontend/src/static/components/icons/eye-icon.jsx
