+++ Frontend/src/static/components/icons/eye-slash-icon.jsx
