+++ Frontend/src/static/components/icons/lock-icon.jsx
