+++ Frontend/src/static/components/icons/user-icon.jsx
